Implement LSP client code actions
The idea is to define an internal plist for the dictionary (same for
disabled rules and false positives) `lsp-ltex--stored-dictionary` in
which, words will be added when the user apply the LSP action "Add
'word' to dictionary", this plist gets automatically stored under
`lsp-ltex-user-rules-path`, and it is loaded at startup.

To allow backward compatibility, the user configured
`lsp-ltex-dictionary` variable is read and combined with the created
`lsp-ltex--stored-dictionary`, to provide the internal variable
`lsp-ltex--dictionary` which is the actual data sent to LTeX-LS via
`ltex.dictionary`.

Fixes: #4 and #16
Ref: https://valentjn.github.io/ltex/ltex-ls/server-usage.html#_ltexaddtodictionary-client